Add counter and registry structs
Still a clunky as I fight with the borrow checker, but does compile.

The counter reference seems to be keeping an implicit mutable reference
to the registry so this won't be usable. I need to move the registry
to keep a vector of reference to counters as we won't be the owners
of the counters really, the application will be.