Get rid of the counter() convenience method
It is still unclear to me how to better implement a method
that returns a reference to a new object. Perhaps using Box is the way
but for now the application needs to explicitly instanciate the counter
and register it in the registry. The registry now has a vector of
immutable references to the counters.